requirements.txt
Pillow==2.7.0
Scrapy==1.0.3
